refactor(fieldmap): first pass moving elements out of fieldmap_react.tsx
